Resolves #128

+++ server/routes/confirmationRouter.js
@@ -1,6 +1,8 @@
 import express from 'express';
 import {ConfirmationService} from "../service/ConfirmationService";
 import {UsuarioService} from "../service/UsuarioService";
+import * as erro from '../util/ErroHandler';
+
 const router = express.Router();
 
 router.get('/:uuid', async (req, res) => {
@@ -8,11 +10,20 @@ router.get('/:uuid', async (req, res) => {
   try {
     const token = await ConfirmationService.retrieveRegistrationToken(uuid);
     let usuario = await UsuarioService.consultaUsuario(token.ownerUsername);
-    const usuarioNovo = {active: true};
 
-    await UsuarioService.editaUsuario(usuario.username, usuarioNovo);
-    await ConfirmationService.deleteToken(uuid);
-    res.status(200).json(token);
+    let usuariosAtivos = await UsuarioService.retrieveActivedUserByEmail(usuario.email);
+
+    if (typeof usuariosAtivos !== 'undefined' && usuariosAtivos.length > 0) {
+      throw Error(erro.CONFIRMATION.CONFIRMATION_ERROR_EMAIL); 
+       
+    } else {      
+      const usuarioNovo = {active: true};
+      await UsuarioService.editaUsuario(usuario.username, usuarioNovo);
+      await ConfirmationService.deleteToken(uuid);      
+      
+      res.status(200).json(token);
+    }
+
   } catch(err) {
     res.status(400).json(err.message);
   }
